+++ src/components/Editor.tsx
@@ -1,6 +1,6 @@
 //@ts-ignore
 import ImageResize from "@looop/quill-image-resize-module-react";
-import { useMemo } from "react";
+import { ChangeEvent, FormEvent, useMemo } from "react";
 // @ts-ignore
 import ReactQuill, { Quill } from "react-quill";
 import { authService, storageService } from "../utils/firebase";
@@ -8,24 +8,14 @@ import { v4 as uuid } from "uuid";
 import { useState } from "react";
 import { useEffect } from "react";
 import { useRef } from "react";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faCircleNotch } from "@fortawesome/free-solid-svg-icons";
+import { EditorTypes } from "../types/Editor.types";
 
-export const Editor: React.FC<{
-  value: string;
-  setValue: React.Dispatch<React.SetStateAction<string>>;
-}> = ({ setValue, value }) => {
+export const Editor: React.FC<EditorTypes> = ({ setValue, value }) => {
   Quill.register("modules/imageResize", ImageResize);
   const [stringedFile, setStringedFile] = useState("");
   const quillRef = useRef<ReactQuill>(null);
-
-  const fileToUrl = (file: File) => {
-    const reader = new FileReader();
-    reader.readAsDataURL(file);
-    reader.onload = (e: any) => {
-      if (e.target?.result) {
-        setStringedFile(e.target.result);
-      }
-    };
-  };
 
   const imageHandler = () => {
     // 파일을 업로드 하기 위한 input 태그 생성
@@ -49,7 +39,7 @@ export const Editor: React.FC<{
     () => ({
       toolbar: {
         container: [
-          [{ header: [1, 2, false] }],
+          [{ header: [false] }],
           ["bold", "italic", "underline", "strike", "blockquote"],
           [
             { list: "ordered" },
@@ -97,39 +87,51 @@ export const Editor: React.FC<{
     "background",
   ];
 
-  console.log(value);
+  const fileToUrl = (file: File) => {
+    const reader = new FileReader();
+    reader.readAsDataURL(file);
+    reader.onload = (e: any) => {
+      if (e.target?.result) {
+        setStringedFile(e.target.result);
+      }
+    };
+  };
+
+  const UploadToFirebase = async () => {
+    const fireBaseFileRef = storageService
+      .ref()
+      .child(`${authService.currentUser?.uid || uuid()}/${uuid()}`);
+    try {
+      const uploadTask = await fireBaseFileRef.putString(
+        stringedFile,
+        "data_url"
+      );
+      const downloadURL = await uploadTask.ref.getDownloadURL();
+      if (downloadURL !== "") {
+        const range = quillRef.current?.getEditor().getSelection()?.index;
+
+        if (range !== undefined && range !== null) {
+          let quill = quillRef.current?.getEditor();
+          quill?.setSelection(range, 1);
+
+          quill?.clipboard.dangerouslyPasteHTML(
+            range,
+            `<img src=${downloadURL}  alt="이미지 태그 입니다." className="w-full"/>`
+          );
+        }
+      }
+    } catch (error) {
+      console.log(error);
+    }
+  };
 
   useEffect(() => {
     if (stringedFile !== "") {
-      (async () => {
-        const fireBaseFileRef = storageService
-          .ref()
-          .child(`${authService.currentUser?.uid || uuid()}/${uuid()}`);
-        try {
-          const uploadTask = await fireBaseFileRef.putString(
-            stringedFile,
-            "data_url"
-          );
-          const downloadURL = await uploadTask.ref.getDownloadURL();
-          if (Boolean(downloadURL)) {
-            const range = quillRef.current?.getEditor().getSelection()?.index;
-
-            if (range !== undefined && range !== null) {
-              let quill = quillRef.current?.getEditor();
-              quill?.setSelection(range, 1);
-
-              quill?.clipboard.dangerouslyPasteHTML(
-                range,
-                `<img src=${downloadURL} alt="imgTag"/>`
-              );
-            }
-          }
-        } catch (error) {
-          console.log(error);
-        }
-      })();
+      UploadToFirebase();
     }
   }, [stringedFile]);
+
+  console.log(value);
 
   return (
     <ReactQuill
