+++ src/components/router.tsx
@@ -3,6 +3,7 @@ import { useEffect } from "react";
 import { useState } from "react";
 import { BrowserRouter, Redirect, Route, Switch } from "react-router-dom";
 import { Campus } from "../pages/Campus";
+import { CampusDetail } from "../pages/CampusDetail";
 import { Forum } from "../pages/Forum";
 import { Home } from "../pages/Home";
 import { UserObjTypes } from "../types/UserObj.types";
@@ -46,6 +47,9 @@ export const Router: React.FC = () => {
         </Route>
         <Route path={routes.forum}>
           <Forum />
+        </Route>
+        <Route path={routes.campusDetail()}>
+          <CampusDetail />
         </Route>
         <Redirect to={routes.home}>
           <Home />
