+++ src/types/CampusDetail.types.ts
@@ -1,0 +1,3 @@
+export interface CampusDetailUseParamsTypes {
+  campus: string;
+}
