+++ src/types/Editor.types.ts
@@ -1,0 +1,4 @@
+export interface EditorTypes {
+  value: string;
+  setValue: React.Dispatch<React.SetStateAction<string>>;
+}
