+++ src/types/PopUpLogin.types.ts
@@ -1,0 +1,4 @@
+export interface PopUpLoginTypes {
+  popUpLoginMode: boolean;
+  setPopUpLoginMode: React.Dispatch<React.SetStateAction<boolean>>;
+}
